feat: split main context menu and its items/submenus for reusability, replace old menu with RowContextMenu and TableContextMenu
